unix: Cast a variable to int from time()
The value from time() is used as a random seed, but Coverity complains
about it being assigned to an int.  Add a cast to silence it.

+++ unix/test_heap.c
@@ -104,7 +104,7 @@ main(int argc, char *argv[])
     if (i < argc) {
 	random_seed = atoi(argv[i]);
     } else {
-	random_seed = time(NULL);
+	random_seed = (int) time(NULL);
     }
     if (debug)
 	printf("Random seed is %d\n", random_seed);
